pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wt-project-base"
version = "0.1.0"
description = "Base project type plugin for wt-tools — universal project knowledge"
requires-python = ">=3.9"
dependencies = ["pyyaml>=5.0"]

[project.scripts]
wt-project-base = "wt_project_base.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.entry-points."wt_tools.project_types"]
base = "wt_project_base:BaseProjectType"

[tool.pytest.ini_options]
testpaths = ["tests"]
